refactor: move Pty into pty/index.ts with self-reexport
